requirements.txt
pytest-homeassistant-custom-component
hass_nabucasa # cloud component
aiohttp_cors # http component
dateparser # tests